\name{tkrgl-package}
\alias{tkrgl-package}
\alias{tkrgl}
\docType{package}
\title{
\packageTitle{tkrgl}
}
\description{
\packageDescription{tkrgl}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{tkrgl}
\packageIndices{tkrgl}


History:
\tabular{ll}{
0.2-2 \tab First public release \cr
0.3   \tab Added possibility to control multiple windows \cr
0.4   \tab Compatibility with 2.0.0 tcltk package \cr
0.5   \tab Added continuous rotation \cr
0.6   \tab Added par3dsave \cr
0.7   \tab Added parameters to \code{\link{spinControl}}, fixed startup \cr
0.8   \tab Minor fixes to pass checks \cr
}
}
